# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for order fulfillment tracking.",
        "title": "Order Fulfillment Tracking API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
        "/api/v1/order-events/import": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Process a batch of order status update events concurrently. Each event is validated and processed in its own DB transaction. The response always returns aggregated counts, even on partial failure.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Event"
                ],
                "summary": "Import batch order events",
                "parameters": [
                    {
                        "description": "List of events",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ImportOrderEventRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ImportOrderEventsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBadReqResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ErrorInternalServerErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ImportOrderEventsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/orders": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Get all orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order Status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Customer Name",
                        "name": "customer_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Ordered Date",
                        "name": "ordered_at",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.OrderReponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBadReqResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorUnauthenticatedResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorUnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "Order details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBadReqResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorUnauthenticatedResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorUnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve detail of an order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Get order detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.OrderReponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBadReqResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorUnauthenticatedResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorUnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{id}/status": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Update order status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBadReqResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorUnauthenticatedResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorUnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.EventError": {
            "type": "object",
            "properties": {
                "order_id": {
                    "type": "integer",
                    "example": 100157
                },
                "reason": {
                    "type": "string",
                    "example": "Invalid transition from 'created' to 'delivered'"
                },
                "status": {
                    "type": "string",
                    "example": "delivered"
                }
            }
        },
...
        "dto.ImportOrderEventRequest": {
            "type": "object",
            "properties": {
                "event_at": {
                    "type": "string",
                    "example": "2026-05-17T09:00:00Z"
                },
                "order_id": {
                    "type": "integer",
                    "example": 100156
                },
                "status": {
                    "type": "string",
                    "example": "packed"
                },
                "updated_by": {
                    "type": "string",
                    "example": "warehouse_staff_01"
                }
            }
        },
        "dto.ImportOrderEventsResponse": {
            "type": "object",
            "properties": {
                "accepted_count": {
                    "type": "integer",
                    "example": 8
                },
                "duplicate_count": {
                    "type": "integer",
                    "example": 1
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.EventError"
                    }
                },
                "rejected_count": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.OrderReponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "ordered_at": {
                    "type": "string",
                    "example": "2026-05-01T00:00:00Z"
                },
                "shipping_address": {
                    "type": "string",
                    "example": "Viet Nam"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.OrderStatus"
                        }
                    ],
                    "example": "paid"
                },
                "total_amount": {
                    "type": "integer",
                    "example": 100000
                },
                "user_phone": {
                    "type": "string",
                    "example": "0123456789"
                },
                "username": {
                    "type": "string",
                    "example": "Supper Man"
                }
            }
        },
        "dto.OrderRequest": {
            "type": "object",
            "required": [
                "shipping_address",
                "total_amount",
                "user_phone",
                "username"
            ],
            "properties": {
                "shipping_address": {
                    "type": "string",
                    "example": "Viet Nam"
                },
                "total_amount": {
                    "type": "integer",
                    "example": 100000
                },
                "user_phone": {
                    "type": "string",
                    "example": "10123456789"
                },
                "username": {
                    "type": "string",
                    "example": "Supper Man"
                }
            }
        },
        "dto.UpdateStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.OrderStatus"
                        }
                    ],
                    "example": "delivered"
                }
            }
        },
        "models.OrderStatus": {
            "type": "string",
            "enum": [
                "created",
                "paid",
                "packed",
                "shipped",
                "delivered",
                "cancelled",
                "refunded"
            ],
            "x-enum-varnames": [
                "ORDER_STATUS_CREATED",
                "ORDER_STATUS_PAID",
                "ORDER_STATUS_PACKED",
                "ORDER_STATUS_SHIPPED",
                "ORDER_STATUS_DELIVERED",
                "ORDER_STATUS_CANCELLED",
                "ORDER_STATUS_REFUNDED"
            ]
        },
        "response.ErrorBadReqResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Invalid input data"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "response.ErrorInternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Internal server error"
                },
                "status": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
        "response.ErrorNotFoundResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Resource not found"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
        "response.ErrorUnauthenticatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Authentication required"
                },
                "status": {
                    "type": "integer",
                    "example": 401
                }
            }
        },
        "response.ErrorUnauthorizedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Permission denied"
                },
                "status": {
                    "type": "integer",
                    "example": 403
                }
            }
        },
        "response.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "pagination": {
                    "$ref": "#/definitions/response.Pagination"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "response.Pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer",
                    "example": 1
                },
                "limit_items": {
                    "type": "integer",
                    "example": 10
                },
                "total_items": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "response.ResponseStruct": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}